bugfix - endpoint de relatórios gerando 403 no front

+++ src/main/java/br/com/utfpr/gerenciamento/server/model/dashboards/DashboardEmprestimoCountRange.java
@@ -17,9 +17,9 @@ public class DashboardEmprestimoCountRange {
   // Construtor adicional para aceitar Long (retorno de COUNT/SUM do JPA)
   public DashboardEmprestimoCountRange(
       Long total, Long emAndamento, Long emAtraso, Long finalizado) {
-    this.total = total != null ? total.intValue() : 0;
-    this.emAndamento = emAndamento != null ? emAndamento.intValue() : 0;
-    this.emAtraso = emAtraso != null ? emAtraso.intValue() : 0;
-    this.finalizado = finalizado != null ? finalizado.intValue() : 0;
+    this.total = total != null ? Math.toIntExact(total) : 0;
+    this.emAndamento = emAndamento != null ? Math.toIntExact(emAndamento) : 0;
+    this.emAtraso = emAtraso != null ? Math.toIntExact(emAtraso) : 0;
+    this.finalizado = finalizado != null ? Math.toIntExact(finalizado) : 0;
   }
 }
